+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.12.3
+
+* added `tags_factory` to `trace_function`